Move provider wait methods to package provider.
